# backend/swagger (11).yaml
swagger: '2.0'
info:
  description: >
    This is a sample server Petstore server.




    [Learn about Swagger](http://swagger.io) or join the IRC channel `#swagger`
    on irc.freenode.net.





    For this sample, you can use the api key `special-key` to test the
    authorization filters
  version: 1.0.0
  title: vitaSpider API Document
  termsOfService: 'http://helloreverb.com/terms/'
  contact:
    name: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: petstore.swagger.io
basePath: /v2
schemes:
  - http
paths:
  /account:
    post:
      tags:
        - account
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/Account'
      responses:
        '201':
          description: > 
                    code: 2000, message: Account successfully created
          schema:
            type: object
            properties:
              meta:
                $ref: '#/definitions/Meta'
              data:
                $ref: '#/definitions/NewUserCreated'
        '200':
          description: | 
                    code: 4001, message: account already existed (activated)
                    
                    code: 4002, message: email is invalid
                    
                    code: 4003, message: password doesn't meet min requirement
          examples: 
            application/json:
              name: Puma
              type: Dog
              color: Black
              gender: Female
              breed: Mixed
              
          schema:
            type: object
            properties:
              meta:
                $ref: '#/definitions/Meta'
              data:
                $ref: '#/definitions/NewUserCreated'
        '400':
          description: 'invalid api call'
          schema:
            type: string
            title: 'bad request'
    put:
      tags:
        - account
      description: >
        Update account user info, not all attributes have to be present. If email
        is being updated, email has to check uniqueness and send activation mail.
        If user email is changed, this account will be logout.
      parameters:
        - in: body
          name: user basic info
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: |
            code: 2000, message: Update success
            
            code: 4000, message: Password doesn't meet requirement
            code: 4001, message: account already exist
            code: 4002, message: account name  invalid
          schema:
            type: object
            properties:
              meta:
                $ref: '#/definitions/Meta'
              data:
                $ref: '#/definitions/User'
        '401':
          description: User password does not meet requirement
          schema:
            type: string
            title: invalid request
...
definitions:
  Account:
    description: same two attributes in User object
    type: object
    properties:
      email:
        type: string
      password:
        type: string
  User:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
      first:
        type: string
      last:
        type: string
      role:
        type: string
  NewUserCreated:
    type: object
    properties:
      id:
        type: string
      email:
        type: string
...
  Meta:
    type: object
    properties:
      code:
        type: number
      message:
        type: string
...